fix: union webpack-scaffold and generators package
