+++ packages/generators/src/init-generator.ts
@@ -2,7 +2,7 @@ import { blue, green, bold } from 'colorette';
 import { utils } from 'webpack-cli';
 import logSymbols from 'log-symbols';
 import path from 'path';
-import { Confirm, Input, List } from '@webpack-cli/webpack-scaffold';
+import { Confirm, Input, List } from './utils/scaffold-utils';
 
 import {
     getDefaultOptimization,
