--- packages/webpack-scaffold/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@webpack-cli/webpack-scaffold",
-  "version": "1.0.3",
-  "description": "Utility files for webpack-scaffold",
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
-  "publishConfig": {
-    "access": "public"
-  },
-  "license": "MIT",
-  "files": [
-    "lib"
-  ],
-  "dependencies": {
-    "jscodeshift": "^0.11.0",
-    "yeoman-generator": "^4.12.0"
-  },
-  "devDependencies": {
-    "@types/yeoman-generator": "^4.11.3"
-  },
-  "gitHead": "fb50f766851f500ca12867a2aa9de81fa6e368f9"
-}
